fix: key bindings in plain text view
