+++ internal/pkg/widgets/plain.go
@@ -1,15 +1,19 @@
 package widgets
 
 import (
+	"github.com/charmbracelet/bubbles/key"
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/muesli/reflow/wordwrap"
+
+	"github.com/hedhyw/json-log-viewer/internal/keymap"
 )
 
 // PlainLogModel is a widget that shows multiline text in a viewport.
 type PlainLogModel struct {
 	viewport viewport.Model
 	text     string
+	keyMap   keymap.KeyMap
 }
 
 // NewPlainLogModel initializes a new PlainLogModel with the given text.
@@ -17,10 +21,12 @@ type PlainLogModel struct {
 func NewPlainLogModel(
 	text string,
 	windowSize tea.WindowSizeMsg,
+	keyMap keymap.KeyMap,
 ) (PlainLogModel, tea.Cmd) {
 	m := PlainLogModel{
 		text:     text,
 		viewport: viewport.New(windowSize.Width, windowSize.Height),
+		keyMap:   keyMap,
 	}
 
 	m = m.refreshText(windowSize.Width)
@@ -47,6 +53,12 @@ func (m PlainLogModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		m.viewport.Width = msg.Width
 		m.viewport.Height = msg.Height
 		m = m.refreshText(msg.Width)
+	case tea.KeyMsg:
+		if key.Matches(msg, m.keyMap.Exit) ||
+			key.Matches(msg, m.keyMap.Back) ||
+			key.Matches(msg, m.keyMap.Open) {
+			return m, tea.Quit
+		}
 	}
 
 	var cmd tea.Cmd
